backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Werkzeug==3.0.1
gunicorn==21.2.0
